better BoundsError message and fix 32bit test failure

+++ test/core.jl
@@ -2038,7 +2038,7 @@ f9534e(x) = (a=IOBuffer(); a.(x) = 3)
 f9534f() = (a=IOBuffer(); a.(-2))
 f9534f(x) = (a=IOBuffer(); a.(x))
 @test try; f9534f() catch ex; isa((ex::BoundsError).a,IOBuffer) && ex.i == -2; end
-@test try; f9534f(typemin(Int64)+2) catch ex; isa((ex::BoundsError).a,IOBuffer) && ex.i == typemin(Int64)+2; end
+@test try; f9534f(typemin(Int)+2) catch ex; isa((ex::BoundsError).a,IOBuffer) && ex.i == typemin(Int)+2; end
 x9634 = 3
 @test try; getfield(1+2im, x9634); catch ex; (ex::BoundsError).a === 1+2im && ex.i == 3; end
 @test try; error(BoundsError()) catch ex; !isdefined((ex::BoundsError), :a) && !isdefined((ex::BoundsError), :i); end
